refactor: rename client/_mesh.py to _mesh_url.py; scaffold mesh view module
The URL resolver lived in client/_mesh.py; rename it to _mesh_url.py to match its contents (resolve_mesh_url / CALFKIT_MESH_URL) and free the 'mesh' name for the new caller-side mesh view. Updates the 3 production import sites (caller.py, _broker.py, cli/_run.py) + a comment, and moves its tests to test_mesh_url.py.

Scaffolds the (empty) calfkit/client/mesh.py + tests/test_mesh.py for the feature built up in the following commits. No behavior change.

Refs: docs/designs/caller-side-mesh-view-impl-plan.md (Commit 0)
